requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
Flask==1.0.2
idna==2.7
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pytz==2018.5
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
whitenoise==4.1
